pyproject.toml
[tool.poetry]
name = "personal-blog"
version = "0.1.0"
description = "My personal blog, managed with Hugo and CDK"
authors = ["Mark Bluemer <mark@theblue.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
"aws-cdk.core" = "^1.71.0"
"aws-cdk.aws-s3" = "^1.71.0"
"aws-cdk.aws-certificatemanager" = "^1.71.0"
"aws-cdk.aws-cloudfront" = "^1.71.0"
"aws-cdk.aws-s3-deployment" = "^1.71.0"
"aws-cdk.aws-route53" = "^1.71.0"
"aws-cdk.aws-route53-targets" = "^1.71.0"
"aws-cdk.aws-lambda" = "^1.71.0"
"aws-cdk.aws-lambda-python" = "^1.71.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
